refactor(identification): Move configuration parsing functions to config module for better organization
